tast-tests: Wait for Guest process to start to identify correct PID
The current way to identify if a Chrome renderer is available does not
work for Guest processes. The method is to wait for the cryptohome user
mount then to get the PID but it seems this is not sufficient when
logging into a Guest session.

This introduces a new method which ensures the browserwatcher.NewWatcher
is instantiated only when the underlying PID that it is watching has
successfully been identified. It also introduces a new method on the
*Watcher instance called WaitExit which ensures the PID that is being
watched successfully shuts down.

BUG=chromium:1174939
TEST=tast run <DUT-IP> security.UserFilesGuest
TEST=tast run <DUT-IP> ui.GuestModeExtension

Change-Id: I22a35a2d553e4448d2572fe8794cbb4813f50247
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2682086
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>
Tested-by: Ben Reich <benreich@chromium.org>