tast: Reduce WiFi Tast tests logging verbosity.
The current default configuration of the logging is very verbose. As a
result, the log size becomes too large to inspect on the Stainless web.
Instead, users need to download and decompress the log file to inspect
it. The CL reduces the logging level and tags to more reasonable values.

BUG=b:179802328
TEST=ran the following tests and checked the logs:
wifi.RoamDbus                  [ PASS ]
network.BasicProfileProperties [ PASS ]

Change-Id: I64b162f38a60fbf9b42cebd85fcae71e64750935
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2683519
Tested-by: Arowa Suliman <arowa@chromium.org>
Reviewed-by: Ben Reich <benreich@chromium.org>
Reviewed-by: Patrick Meiring <meiring@google.com>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>
Commit-Queue: Arowa Suliman <arowa@chromium.org>