tast: Reduce WiFi Tast tests logging verbosity.
The current default configuration of the logging is very verbose. As a
result of that the log size has increased which made looking at tests
much harder, because it adds two or three more steps (download, possibly
transfer to a remote workstation, decompress). Reduced the logging level
and tags to a more reasonable values.

BUG=none
TEST=ran the following tests and checked the logs:
wifi.RoamDbus                  [ PASS ]
network.BasicProfileProperties [ PASS ]

Change-Id: I64b162f38a60fbf9b42cebd85fcae71e64750935