tast: Reduce WiFi Tast tests logging verbosity.
The current default configuration of the logging is very verbose. As a
result, the log size becomes too large to inspect on the Stainless web.
Instead, users need to download and decompress the log file to inspect
it. The CL reduces the logging level and tags to more reasonable values.

BUG=b:179802328
TEST=ran the following tests and checked the logs:
wifi.RoamDbus                  [ PASS ]
network.BasicProfileProperties [ PASS ]

Change-Id: I64b162f38a60fbf9b42cebd85fcae71e64750935