chrome: Always save minidump snapshots when chrome.New times out
In crbug.com/1172863 we are seeing unknown browser hang. To debug
the issue, we want to take minidump snapshots of browser-related
processes when we hit the timeout.

Actually we already take minidump snapshots when chrome.New fails
on restarting Chrome. This patch expands the coverage to all cases
where chrome.New times out.

BUG=chromium:1172863
TEST=tast run DUT ui.ChromeLogin  # w/ chrome.LoginTimeout mod

Change-Id: Ifc9a5ffd0b1f6cabdbb17f73f279e9cefaeaf244