[lacros Tast] Add LacrosExtraArgs
This is used in a follow-up CL. If we want to specify extra args to lacros,
we should make sure they apply no matter whether lacros is launched via
the shelf or command line. So, add LacrosExtraArgs to chrome.Config
where it is accessible.

BUG=chromium:1140407
TEST=fast_build.sh -T

Change-Id: Ic320045be7b2fa50a84954b6724d764cdb4bbccf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2684034
Commit-Queue: Eliot Courtney <edcourtney@chromium.org>
Tested-by: Eliot Courtney <edcourtney@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>