tast: Add ArcConnectivityTest app
The new app allows a tast test to connect to a remote target from an
ARC app through the PlayStore global proxy and report the response.

This app is used in a follow up CL to test the integration with the
Chrome OS system-proxy daemon, which does proxy authentication on
behalf of ARC apps.

BUG=b:179141883
TEST=emerge and deploy tast-local-apks-cros

Change-Id: Icbddf36a17fe46d42a697f8cd38e2ddb15775ffc