Move code for enabling high vis in the UI to common package.
On the road to a CB->CB test. I will start with a a smoke test that turns
on the high vis mode on each chromebook. First step it to extract that
logic to a package. This CL will also help us

1. Settle on a naming scheme for API vs UI interactions.
2. Making serviceizing code for remote tests simple: tests and services
will call the same code instead of duplicating it.

BUG=chromium:1176399
TEST=nearbyshare.HighVisibilityUISmoke

Change-Id: I19d7275a9093851fa00ae6534ac78f29b2f11ef8