Add Nearby Share service.
For remote tests (e.g CB->CB), we need to call services on the DUT to
have it execute local code. This CL creates the service with some
initial functions that let us login, setup nearby share, and enable high
vis mode with the UI.

nearby_share_service.proto is the api.
nearby_share_serivce.go is the implementation.
The rest are boiler plate or autogenerated.

BUG=chromium:1176399
TEST=nearbyshare.HighVisibilityUISmoke

Cq-Depend: chromium:2688004
Change-Id: Ia70436011b9df5a4ef501446043dc790b5b619ed