tast-tests: Remove precondition and use uiauto for filemanager.DrivefsUI
Majority of the flakes for filemanager.DrivefsUI are due to the
precondition. The flaky part of the precondition can be bypassed and the
main use of it is not required in this test.

This CL also ports the test to the new uiauto library to improve
readability.

BUG=chromium:1177053
TEST=tast run <DUP-IP> filemanager.DrivefsUI

Change-Id: Id5fa582343792804fa70ef932b692bc51bfd2b82
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2690062
Reviewed-by: Austin Tankiang <austinct@chromium.org>
Reviewed-by: Brendan Hansknecht <bhansknecht@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>
Tested-by: Ben Reich <benreich@chromium.org>