(Nearby Share) Add test for in-contacts Android->CrOS transfer
Add a UI-based test for Android->CrOS transfer. Sending on the Android
side is accomplished the same as in the high-visibility tests, using the
Nearby Snippet APK. Receiving on CrOS is controlled through the UI by
interacting with the in-contact receiving notifications. Contacts setup is achieved by manually signing in the Android phone to an infinite-password OTA, while the CrOS OTA sign-in is achieved with internal credentials. The 2 OTAs have been manually configured to be mutual contacts.

BUG=chromium:1165773
TEST=tast run <dut>
nearbyshare.AndroidSenderCrosReceiverInContacts.small_jpg
TEST=tast run -var=rooted=false <dut>
nearbyshare.AndroidSenderCrosReceiverInContacts.small_jpg

Cq-Depend: chrome-internal:3607770
Change-Id: I257f580f61897895aad5e59de5e2c0223705a831