Add Fixture (precondition v2) for Nearby Share local tests.
This fixture will share the setup of the ChromeOS side in Nearby share
tests. This will make test runs faster for eng and in the lab.

Before to run 6 tests: 3m 06s
After to run 6 tests: 2m 09s

The fixture could not read the nearbyshare/* libraries in
local/bundles/cros/ so they had to be moved to the chrome/nearbyshare
directory. Moving these folders caused import cycles so I have moved constants
around to be with the packages that use them.

Fixtures that call chrome.Lock() need to be added to a dictionary in
chrome.go.

BUG=chromium:1159975
TEST=nearbyshare.* all passing locally and in the lab

Change-Id: I52f0ec240da0805948fa639599ce2e6b46ebc26a