Tast: Add hwdep WifiIntel to tests
Several tests are currently using the softwaredep "intel_wifi_chip" in
combination with hwdep SkipOnModel to exclude devices with non-intel
WiFi. This functionality has been added as a single hardwaredep
"WifiIntel", so this CL replaces the instances of that usage with the
WifiIntel hwdep. It also updates relevant comments and TODO's to be
more up-to-date.

BUG=b:180133528
TEST=tast run ${DUT} ${TEST}
Should pass on DUT with Intel WiFi, and skip on DUT without Intel Wifi
Repeat for each affected test

Cq-Depend: chromium:2689451
Change-Id: I61934fdbcd3d8148321859d72c9d4b7fa07beb9a