scanner: Stop ippusb-bridge job when needed
For simulating a network scanner, we need to run ippusb_bridge with
specific parameters.  Because udev manages ippusb_bridge by default, we
need to stop that job before trying to start it manually.  This doesn't
call upstart.StopJob because it needs to pass additional environment
variables in order to stop the correct ippusb-bridge job instance.

BUG=b:179413348
TEST=tast run printer.* scanner.*

Cq-Depend: chromium:2689607
Change-Id: Ibb3686162871e9f37264e668b55ed43dad94e0cf