cros: Add logger to init script.
This handles the fact that /usr/bin/logger is used in many init conf
script. In very most cases its invocation is short that has little
chance to be visible from SELinux tests. However problem exists. In some
case, logger is bound to longer command and it becomes visible by
SELinux tests with higher chance.
This standardize this situation and add logger as allowed call from init
scripts.

BUG=chromium:1175872
TEST=Modified arc-ureadahead.conf by adding sleep in order to make sure
it is running when body of the test is executed and could repro failures
in 100% cases. With this fix this behavior is gone.

Change-Id: I6d3dd74842a049364ae1a625ac3505fc71befee3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2691724
Tested-by: Yury Khmel <khmel@chromium.org>
Reviewed-by: Qijiang Fan <fqj@google.com>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Satoshi Niwa <niwa@chromium.org>
Reviewed-by: Patrick Meiring <meiring@google.com>
Commit-Queue: Yury Khmel <khmel@chromium.org>