tast-tests: Rename arc.BuiltInCrash.* to arc.CxxCrash.*
This Tast test is originally arc.NativeCrash because it is
for "native_crash" of Android, but it's mechanically renamed
at commit 782757c70a8ec8a3aca9ed624f9fcb70c3ad6c26.

However, "built in" is not a so good phrase to describe crashes
tagged as "native_crash".  Many Java binaries are "built in"
in Android.  The phrase is confusing.  Also "system" is a
widely used replacement of "native", but this word is already
reserved in Android for another meaning.

The characteristics of "native_crash" of Android is that they
are not crashes of Java programs.  Therefore, I think "non-Java"
crashes or "C++" crashes are appropriate for the replacement.

This replacement is provisional. We want to follow when
Anrdoid renames the "native_crash" to some other string.

BUG=b:169638371
TEST=tast run my_crbook arc.CxxCrash.{mock_consent,real_consent}

Change-Id: Icc1db03e0bdb92bd37a35fe467ae5fd0540c10c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2690772
Tested-by: Kimiyuki Onaka <kimiyuki@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Kimiyuki Onaka <kimiyuki@google.com>