Update BookmarkBarEnabled test
A new feature is being introduced in the bookmarks which modifies
the workflow when star button is clicked to bookmark a page.
Currently, the behaviour is to show a 'Bookmark Added' popup
but according to the new behaviour, a different popup is shown.

This test requires adding a new bookmark which it is currently
doing by clicking on the star button. We could either enable
the feature in the test and update it according to the new workflow
or just use keyboard shortcut to add a new bookmark.
This CL adapts the keyboard shortcut approach as it seems to be
simpler.

BUG=chromium:1177671
TEST=tast run <ip> policy.BookmarkBarEnabled

Change-Id: Ie02e5b258b0dd64ebe37c356fc24e78e1364cd99