tast-tests: Add arc.DownloadManager test
The test checks whether ARC can download files through Android's
DownloadManager with Android apps.

The overall structure of the test is as follows:

1. Start a local HTTP server that serves a test file on the CrOS side.
2. Download the test file from the Android side using an Android app.
  2.1. Set up reverse port forwarding so that the local server on the
       CrOS side can be accessed from the Android side.
  2.2. Install a new APK ArcDownloadManagerTest.apk.
  2.3. Start the main activity of ArcDownloadManagerTest. It downloads
       the test file to Android's Download/ using DownloadManager.
3. Read the downloaded file from the CrOS Downloads/ (which should have
   the same contents as Android's Download/), and check whether it
   matches the original test file.

The reverse port forwarding in 2.1. is achieved by the command "adb
reverse", which is available via two new functions ReverseTCP (to start
the forwarding) and RemoveReverseTCP (to stop the forwarding) added to
chromiumos/tast/local/android/adb and chromiumos/tast/local/arc/commands
(the latter is a wrapper of the former).

BUG=b:180082486
TEST=$ emerge-${BOARD} tast-local-apks-cros
TEST=$ cros deploy --root=/usr/local --board=${BOARD} <DUT> \
TEST=  tast-local-apks-cros
TEST=$ tast run <betty/betty-arc-r> arc.DownloadManager*

Change-Id: Ib36f5113bfb42ce83625aeaf72831301eb91304b