(Nearby Share) CB->Phone in-contacts test
Add an in-contacts version of the CB->Phone test. Android's Nearby
Snippet doesn't support in-contacts sharing yet, so receiving on the
Android is controlled using UIAutomator. For in-contacts sharing, a
half-screen card will appear on the Android's screen once the CB selects
it as a receiver. The test waits for this card to appear by checking for
its title, which contains the sender's device name. Then we click the
"Accept" button and wait for the card to be gone, since it disappears
once the share finishes.

BUG=chromium:1165773
TEST=tast run <dut>
nearbyshare.CrosSenderAndroidReceiverInContacts.small_jpg

Change-Id: I4c41337a5e75d61a280925a9b233c2fb991860db