tast-tests: typec: Make ModeSwitch nominally runnable on CQ
While typec.ModeSwitch needs a Thunderbolt(TBT) device to execute,
there is value in at having it at least compile and run on
CQ (where the machines don't have connected TBT devices).

So, add some HWDeps for a couple of boards which support the new TCPMv2
ectool commands, and add a check right at the beginning of the test
which goes through the PD identity data of the partner to look for a
connected Thunderbolt device (and returns early if one is not present).

BUG=b:180047526
TEST=Make sure typec.ModeSwitch still runs OK in both the following
     scenarios:
     - TBT device connected.
     - No TBT device connected.

Change-Id: I2f6d54c386e325ad601cbe5740da860fbbfaf282
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2699451
Reviewed-by: Benson Leung <bleung@google.com>
Reviewed-by: Jun Sheng <shengjun@chromium.org>
Tested-by: Prashant Malani <pmalani@chromium.org>
Commit-Queue: Prashant Malani <pmalani@chromium.org>