scanner: Add StartVirtualScanner test
When working on scanning changes, it is useful to have the virtual USB
scanner available for rapid testing.  Normally the scanner.* tests set
up the device and tear it down again, so you have to re-run the whole
test repeatedly if you want to iterate on a change.

scanner.StartVirtualScanner provides a manual subset of the main tests
similar to dev.RemoteDesktop where it sets up the virtual USB devices
and then ends.

BUG=None
TEST=tast run scanner.StartVirtualScanner

Change-Id: I9cd5c0704b1e17cc44aad6f0e409afef2808d75e