wifi.HostapHwsim.full: increase timeout
NB: this 'full' variant doesn't run automatically anywhere; it's
currently for local testing.

I've found that depending on how you're running this -- e.g., with extra
instrumentation, or on slower hardware (it currently runs only on VMs,
but I run it on ARM DUTs for some cases) -- it can take on the order of
an hour. Bump this up.

BUG=none
TEST=run it

Change-Id: Ice3b886e94c6b69298973a669a3ccd0432997372