Fix policy.Homepage flaky test
This test requires navigating to the homepage URL which it is currently
doing through keyboard hotkeys.
The test is flaky only on a few boards as per the dashboard.
Please find the link to the logs in the bug.

The screenshot in the failure logs show chrome://version page which is
the initial page with which we start the test. It looks like that the
test is not able to navigate to the homepage.This CL updates the method
to do so by clicking on the home button in the toolbar instead of using
keyboard shortcut.

BUG=chromium:1153639
TEST=tast run <ip> policy.Homepage.*

Change-Id: I039e7653715d4f65d0e3b2ec063c4da153315942