wifi tests: add WiFi oncall as primary owner
Rationale: "stable" (i.e., not "in development"; ideally, in a blocking
suite like group:mainline) WiFi tests should be owned by the WiFi team
as a whole. We're using chromeos-wifi-champs@ to reach the oncall for
the week.

Unstable (with 'wificell_unstale' or 'informational' attributes) are
still the responsibility of the test writer to stabilize.

Rules I used:

  mainline+critical = owned by oncall
  mainline+informational = primarily owned by test owner, secondarily
    oncall
  wificell_*+wificell_unstable = primarily owned by test owner,
    secondarily oncall
  wificell_* (stable) = owned by oncall

In a few exceptions (e.g., network.HostapHwsim) I've chosen to retain a
secondary owner for stable tests, as the tests are extra peculiar (e.g.,
VM test setups) and the author (in this case, me) is most familiar.

Not yet covered: network.* tests that aren't primarily WiFi tests.

BUG=none
TEST=none

Change-Id: Ic1ba3858e459a76d1da2e782bbe49096da105914
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2705286
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>