tast-tests: Don't wait for charger-thermal to cool down.
A lot of the tests fail on trogdor because they wait for the
charger-thermal to cool down, and it often doesn't in time. Looking at
several test failures I plotted the temperature against time to clearly
show that when the DUT is charging it'll be unable to cool down this
sensor even if we extend the timeout to 10 minutes.

Hopefully the CPU won't thermally throttle from failing to cool down
this sensor.

BUG=b:180696076
TEST=tast run -buildbundle=crosint <DUT> platform.MLBenchmark.smartdim*

Change-Id: I6bdc9b2c74d85b35628a4c31cb0a5b71b89975fd