Fix the class name find parameter for webview.
It seems the class name of webview has been changed from WebView
to ContentsWebView.  This CL follows the renaming, though it's
a bit tricky; current 4414.0 doesn't include this renaming,
but the next uprev (and Chromium ToT) contains this. For now,
it's using a regexp to match both cases. This should be switched
to a simple ClassName parameter once the next uprev happens and
it's being stabilized.

BUG=chromium:1180422
TEST=manually

Change-Id: I1c5d2e8cfddc35cdaf84e0d41c04ee822cb3a12b