Appcompat: Clean up multiple appcompat test
BUG=b:180728870
TEST=tast run <eve> arcappcompat.*

Observed that multiple apps are failing in launch due to new dialog boxes and change in launch verifier ID. Made changes in the code to fix it. Also added a Handle Dialog Box method to handle dialog boxes in the testutil file. By this fix, flaky launch failures could be avoided.

Change-Id: Ib9597e5a3edcbb369d976c4b5773b67a348071fb