tast-tests: Add apps.SystemWebAppsInstallGuest tests
Verify that System Web Apps (SWA) get installed correctly when a guest
login is started by listing the installed SWAs and comparing to the
number of internal names from the SWA manager.

This tests both the enabled apps are installed and (by passing the
EnabledAllSystemWebApps flag) that all SWAs are installed correctly. By
correctly, ensuring Chrome doesn't crash is a sufficient signal.

BUG=chromium:1122877
TEST=tast run <DUT-IP> apps.SystemWebAppsInstallGuest*

Change-Id: Ia6eb8803ff74b91c8a5ff344e6aeb29a127a3b3e