camera.CCAUISmoke.fake: Remove SkipOnPlatform() for scarlet
The SkipOnPlatform() hwdep was used to workaround a unknown crash when
starting Chrome in Chromium infra. We have largely refactored the test
bootstraping code in the camera app during the SWA migration and the
crash might be already fixed.

Note that the other test like example.ChromeExtension that were suffer
from the same crash on scarlet are also quite stable recently.

BUG=b:172343409,b:178057757
TEST=Pass camera.CCAUISmoke.fake 10 times on scarlet

Change-Id: Ia894a317b13ac6992a0e74ffbb0c047f9c5a1edb