tast-tests: Add jacuzzi to skiplist for tast persistence tests
Jacuzzi devices are known to fail in the lab when
rebooted(b/178449023). The
persistence tests reboot the DUT and thus can impact the rest of the
test suite. Until this reboot issue is fixed, we will not run the
persistence tests on jacuzzi devices.

BUG=b:181057823
TEST=tast run $DUT wifi.PersistenceBluetoothSansWifi,
wifi.PersistenceWifiSansBluetooth (on multiple jacuzzi devices)

Cq-Depend: chromium:2716967
Change-Id: Ica2461f061e08049ea5b2a22f8d7a94819d72c65
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2715659
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Billy Zhao <billyzhao@chromium.org>
Reviewed-by: Po-Hsien Wang <pwang@chromium.org>
Commit-Queue: Billy Zhao <billyzhao@chromium.org>
Tested-by: Billy Zhao <billyzhao@chromium.org>