tast: Enable running CDMOEMCrypto test
This adds the SoftwareDeps and Attr to enable running the CDMOEMCrypto
test. It requires the 'protected_content' SW deps, which is linked to
the cdm_factory_daemon USE flag. It's set to run in the graphics
nightly suite.

This also adds a 15 minute timeout because in practice the CDM test
takes just over 2 minutes to run and the OEMCrypto test takes just over
10 minutes. The reason this takes so long is because HW timing is
involved in the tests.

BUG=b:155512423
TEST=tast run DUT video.CDMOEMCrypto.*

Cq-Depend: chromium:2715657, chromium:2715660
Change-Id: I4d330b629532887596d1edce6364db53dfbbc7cd