tast-tests: Increase Sharesheet interval when clicking icons
The default interval for clicking icons on the Sharesheet is too low for
the low-end devices. It appears that the accessibility tree is not being
updated with the new Coords during animation which leads to premature
clicks on the icon causing flakiness.

The current animation interval for the C++ code is 150ms, so increase
the pollOpts to 1 second to ensure it is adequate for lower-end devices.
When testing on some low-end devices I've seen upper bounds of 0.5s to
animate.

BUG=chromium:1176552
TEST=./tools/repeat-test.sh -s -r 100 -- <gnawty> apps.Sharesheet
Disallow-recycled-builds: test-failures

Change-Id: I860e26de248bb3b936c1fb40b3b2ec794e4cbc3a