tast-tests: Zoom the doc to fit in MeetCUJ tests.
At the default zoom level, when the screen is split between Docs and
Meet, the doc does not fit within the window, causing the doc to be
scrolled during typing, potentially reducing performance in a fashion
not reflecting actual user flows. Zoom to fit prior to any typing to
avoid this.

BUG=b:176840989
TEST=ui.MeetCUJ.16p_notes

Change-Id: I6a8a36156ad2bcc0c109e2561d51b62f2ceeae97