wifi.MalformedProbeResp: Ensure bgscan is off in the test.
The test used to depend on the default behavior that bgscan is off when
only one endpoint is found for the connected service. However, it is
always possible that bgscan configuration is not properly restored. For
example, previous task is killed in the middle, or the DUT gets offline
for a while causing tests unable to restore the setting.

We're seeing a few EBUSY errors in "iw scan", so this probably can
happen from time to time. Ensure background scan is off in the test to
reduce possible flakiness.

BUG=b:171763917
TEST=tast run $DUT wifi.MalformedProbeResp

Change-Id: Ife6d889810e618c4e3e49feddad4f6c6643c4dca