hwsec: combine the TPM reset in remote & local tast
Merge the soft-clear and hard-clear code into the hwsec common helper.

BYPASS_INCLUSIVE_LANGUAGE_REASON = This CL is moving the constant paths
that need to be cleared when we are resetting the system state. And the
policy path is widely use in Chrome OS, we couldn't done it in this CL.
BTW, we have another bug(b/179318995) for COIL in hwsec.

BUG=b:181183725
TEST=tast run $DUT hwsec.*

Change-Id: I17e861f9eebacf7ec501b5d27d741981bb66fc81
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2717983
Tested-by: Yi Chou <yich@google.com>
Reviewed-by: Leo Lai <cylai@google.com>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Yi Chou <yich@google.com>