network: Add a gRPC service for testing the CrOS allowlist
For Chrome devices to work on a domain with TLS inspection or on
networks with restricted external traffic, connections to some
hostnames need to bypass the TLS inspection or firewall. These
hostnames are listed on our support page[1].

This CL adds a gRPC service, tast.cros.network.Allowlist, used by remote
tests to verify that the list of hosts from the Google support page is
up to date with every release.

This service allows setting up a firewall on the device and performing
basic actions like logging in and installing ARC apps (using Google
production end points) behind the firewall.

Testing the allowlist is done in remote tests (instead of local tests)
because it requires making changes to the iptable rules. The safest way
to restore the device to the initial state is to reboot the device.

Implements allowlist.TestSystemServicesConnectivity which tests
that tlsdated works behind the firewall.

go/tast-for-allowlist

[1]https://support.google.com/chrome/a/answer/6334001?hl=en

BUG=b:181110031
TEST=tast -verbose run $DUT allowlist.SystemServicesConnectivity

Cq-Depend: 3638083
Change-Id: I50dc94a7968b3444d7b9507c8199b24f176f8c75