Tast-tests: Update unblocked terms.
As part of cleaning up the code base of offending words,
disable unblocked terms that are clean or can not be modified.
see go/tast-coil-status for details.

Add informations to clearly state how to handle blocked terms
pre-submit failures for terms that can not be modified.

BUG=chromium:1101215
TEST=search_blocked_words.sh -i ../../repohooks/blocked_terms.txt -r

Change-Id: I52adab438e8cc7ac5c75615dca2d1bf05335f9d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2717280
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Laurent Chavey <chavey@google.com>
Commit-Queue: Laurent Chavey <chavey@google.com>