crostini: Fix LaunchBrowser Tast test
In the last month or two the LaunchBrowser tests have been failing
when executing the "env" test using the $BROWSER variable. The
failure looks to be related to either a change in libbrillo for
creating/executing processes, or a change to go in its own os/exec.
The effect is to make it look like vsh has become stateful with
args previously passed after "--". Locally, I have made this test
pass if it runs first. This is a temporary fix until we know what
really broke, since this test has been stable for years.

BUG=chromium:1180485
TEST=tast run host:port crostini.LaunchBrowser.dlc_buster_stable

Change-Id: I1e6c9d50008a00d5ec6830ccadcc85ea712447ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2719243
Tested-by: Nicholas Verne <nverne@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Nicholas Verne <nverne@chromium.org>