network: VPNConnect: Fix and simplify IPv6 ping check
Currently when ping exits with error and does not have a valid output,
the check code will also evaluate `res.Received`, and cause a null
pointer dereference.

This patch fixes this issue and also simplify that check: we only need
to check if |res.Received != 0|, and this is available only when err is
not nil.

BUG=b:179882848
TEST=`tast run home-crbook network.VPNConnect.l2tp_ipsec_psk` w/ and w/o
 ipv6

Change-Id: I39c79e54e84ab8d9b3638bfe92c513f15b0b3ce0