crostini: Don't close tabs in LaunchBrowser
LaunchBrowser occasionally fails with an error about being unable to
find the expected browser tab. This seems to happen because we try to
close the tab after checking for it's presence in each test case, and
this can race with opening the tab for the next test case. If this
happens the new tab will open, and then be closed by the close command
from the previous test case, before we can detect it's presence.

There's no actual need to close these tabs, as each test does
cr.ResetState before running to clean up, and not doing so should make
the failure screenshots more useful anyway.

BUG=chromium:1171103
TEST=Ran test in a loop for a while

Disallow-Recycled-Builds: test-failures
Change-Id: I255b98071d0590e86b932336d7864f9199e9c72f