chrome: Support returning errors from Option
In the next CL we're introducing an option that can fail. Thus update
the signature of Option to allow returning an error. For now, all
existing options return success (nil).

This is just a preparation for the next CL, not changing the current
behavior.

BUG=chromium:1182155
TEST=fast_build.sh -T

Change-Id: Ib8308e5778fa3ada2c0cf5ecabda2303a6b2a545
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2719743
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Jun Mukai <mukai@chromium.org>
Reviewed-by: Fergus Dall <sidereal@google.com>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>