tast-test: Wait for the tinyproxy process to start
Currently the go wrapper over tinyproxy will start tinyproxy and
immediately fork, leaving the proxy instance to run in the background.

tinyproxy will write the pid of the main process to the path provided
via the config option `PidFile`. The pid is then provided to patchpanel
so it can setup an isolated network namespace for the process.

With the current implementation, we run into a race condition where
patchpanel will be invoked to setup the net namespace before tinyproxy
writes the pid to the file.

BUG=b:177840593
TEST=tast run $DUT network.TestProxyServer

Change-Id: I24f6b70c07d3f61fd1648c6becfcc89a219ed256