tast-test: Wait for the tinyproxy process to start
Currently the go wrapper over tinyproxy will start tinyproxy and
immediately fork, leaving the proxy instance to run in the background.
This is done by calling minijail0 with the "-i" option.

tinyproxy will write the pid of the main process to the path provided
via the config option `PidFile`. The pid is then provided to patchpanel
so it can setup an isolated network namespace for the process.

With the current implementation, we run into a race condition where
patchpanel will be invoked to setup the net namespace before tinyproxy
writes the pid to the file. We need to wait for tinyproxy to start
before trying to read the pid, instead of existing imediately after
fork.

BUG=b:177840593
TEST=tast run $DUT network.TestProxyServer

Change-Id: I24f6b70c07d3f61fd1648c6becfcc89a219ed256
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2718089
Tested-by: Andreea-Elena Costinas <acostinas@google.com>
Reviewed-by: Patrick Meiring <meiring@google.com>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Andreea-Elena Costinas <acostinas@google.com>