Re-login the Android device at the beginning of Nearby Share tests
OTA logins expire after about a week, so we need to re-login at the
beginning of test runs to avoid doing it manually. We can do this using
the GoogleAccountUtil APK from tradefed (some details here:
go/editors-android#adding-accounts-to-a-device-via-adb). If running the
tests locally on an already-signed-in device, you can specify the new
'androidLoggedIn' runtime var to skip adding the test account to the
phone.

BUG=chromium:1171010
TEST=tast run <lab DUT> nearbyshare.*
TEST=tast run -var=rooted=false -var=androidLoggedIn=true <local DUT w/
nonrooted signed-in Android> nearbyshare.*

Change-Id: I24c0de2a7503cf2e391ba8641d14935817499f7c