wifi.ChannelScanDwellTime: Use channel 36, not 1
Intel's driver has a different behaviour on channel 1, 6 and 11 where it
dwells for 90ms (IWL_SCAN_DWELL_EXTENDED).
Change the test to use channel 36 in 11n-mixed mode instead, which
should be relatively typical.
Instead of 90ms dwell time, the test now reports ~10ms dwell time on
HrP2, which is more in line with expected results.

BUG=b:181294019
TEST=tast run $HRP2_DUT wifi.ChannelScanDwellTime reports
2021/02/25 22:49:19 [22:49:19.907] Scan found 11 APs
2021/02/25 22:49:23 [22:49:23.108] Dwell Time: 10.49ms

Change-Id: I515fd424796a2c293aa72fbc9fa209dd048780af