multivm fincore metrics for crosvm disks
This CL adds metrics for the amount of file memory that is resident for
crosvm disks. It looks for all crosvm processes, extracts the disk paths
from the crosvm command line arguments, and uses fincore to get the
amount of memory that is resident.

The multivm.MemoryMetrics helper method is introduced to make calling
all the memory metrics collection methods from tests easier.

BUG=b:178980260
TEST=tast run dut multivm.Login.arc_crostini multivm.Lifecycle.arc
multivm.LifecycleShifting.arc_host

Change-Id: I10426473f53cbdf0b49f444523825e2e1d5d93ef