BootMode: Copy Tast files to DUT in new boot mode
When Tast starts, certain files are installed onto the DUT. These are
used for running local tests, and for hosting an RPC server on the DUT.

However, when a test reboots into a recovery image mid-test, these files
are not available on the recovery image. As a result, if the test needs
to connect to RPC after switching boot mode, the test will fail.

This CL addresses this problem in the mode-switcher by copying Tast
files from the DUT onto the test-server, and then re-installing them
onto the DUT after rebooting.

BUG=b:179804701
TEST=Run firmware.BootMode.rec on a newly-leased DUT

Change-Id: Ife4a3aaa010d4a9ab21acee3f5973ec7c24e020a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2725432
Tested-by: Greg Edelston <gredelston@google.com>
Commit-Queue: Greg Edelston <gredelston@google.com>
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>