Storing chrome log to the outdir of the test case.
As noted in crbug.com/1174574, sometimes test developers want
to see the chrome log of the specific test case, but it's
not always easily accessible. This CL introduces logsaver
package to split a log file into multiple test cases, and
provides the code of PreTest/PostTest of chrome's fixture
as well as loggedInTOCUJUser fixture for the proof-of-concept.

BUG=chromium:1174574
TEST=tast run on various tests

Change-Id: I87415ed2c6bbaff4947f201905b0239e439f86a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2747194
Tested-by: Jun Mukai <mukai@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>