Split unstable boards for UI performance tests for CQ
ui.DesktopControl and ui.TabletOperations are sometimes flaky
on certain boards, but the goal of those tests are to detect the
significant regressions and provides failures in CQ and chromium
postsubmit bots. For now, the boards used in CQ are mostly
stable, so I think it's reasonable to group the 'unstable'
boards explicitly, and mark them as 'unstable'.

For ui.TabletOperations, it's now stable, so this CL also
promotes it to critical. for ui.DesktopControl, a failure
is seen on kukui/krane. So this is not promoted yet.

Tastboard links:
https://tastboard.corp.google.com/test/ui.DesktopControl
https://tastboard.corp.google.com/test/ui.TabletOperations

BUG=chromium:1173315
TEST=none

Change-Id: I159c7e0bdca23629e9a21bbef660131aa058ac87
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2725620
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>
Tested-by: Jun Mukai <mukai@chromium.org>