Split target boards and promote tests to critical on CQ boards
ui.DesktopControl and ui.TabletOperations are sometimes flaky
on certain boards, but the goal of those tests are to detect the
significant regressions and provides failures in CQ and chromium
postsubmit bots. Because of that, I think the stability on
the devices which are used actually on CQ should be focused.

Based on tastboard, these tests are pretty stable on the boards
used for chromeos-chrome uprevs and chromium postsubmit bots.
For the time being, I think it's a reasonable idea to split the
tests into 'unstable' category, and make critical on the boards
which are actually used in CQ.

BUG=chromium:1173315
TEST=none

Change-Id: I159c7e0bdca23629e9a21bbef660131aa058ac87